Separate eager execution deferred value utils from legacy deferred value utils which relies on deferred nodes instead of deferred tokens
